+++ crates/mds-tui/src/util.rs
@@ -1,7 +1,4 @@
-use ratatui::{
-    Frame,
-    layout::{Constraint, Flex, Layout, Rect},
-};
+use ratatui::layout::{Constraint, Flex, Layout, Rect};
 use tui_textarea::TextArea;
 
 /// Centers a [`Rect`] within another [`Rect`] using the provided [`Constraint`]s.
@@ -23,12 +20,6 @@ pub(crate) fn center(area: Rect, horizontal: Constraint, vertical: Constraint) -
         .areas(area);
     let [area] = Layout::vertical([vertical]).flex(Flex::Center).areas(area);
     area
-}
-
-pub(crate) fn centered_80_percent(frame: &Frame) -> Rect {
-    let horizontal = Constraint::Percentage(80);
-    let vertical = Constraint::Percentage(80);
-    crate::util::center(frame.area(), horizontal, vertical)
 }
 
 pub(crate) fn text_edit_content_len(txt_edit: &TextArea) -> u16 {